extract filter component
